Evolve3DMorph

+++ robot.py
@@ -7,18 +7,21 @@ import pyrosim.pyrosim as pyrosim
 from motor import MOTOR
 from pyrosim.neuralNetwork import NEURAL_NETWORK
 from sensor import SENSOR
-
+import math
 
 class ROBOT:
-    def __init__(self, solutionID):
+    def __init__(self, solutionID, save=False):
         self.solutionID = solutionID
         self.robotId = p.loadURDF(f"body{self.solutionID}.urdf")
-        os.system(f'rm body{self.solutionID}.urdf')
+        if not save:
+            os.system(f'rm body{self.solutionID}.urdf')
         pyrosim.Prepare_To_Simulate(self.robotId)
         self.Prepare_To_Sense()
         self.Prepare_To_Act()
         self.nn = NEURAL_NETWORK(f"brain{self.solutionID}.nndf")
-        os.system(f'rm brain{self.solutionID}.nndf')
+        print(save)
+        if not save:
+            os.system(f'rm brain{self.solutionID}.nndf')
 
     def Prepare_To_Sense(self):
         self.sensors = {}
@@ -47,7 +50,8 @@ class ROBOT:
     def Get_Fitness(self):
         basePositionAndOrientation = p.getBasePositionAndOrientation(self.robotId)
         basePosition = basePositionAndOrientation[0]
-        xPosition = basePosition[0]
+        xPosition = math.inf if math.isnan(basePosition[0]) else basePosition[0]
+        
         with open(f"tmp{self.solutionID}.txt", "w") as f:
             f.write(str(xPosition))
         os.system(f'mv tmp{self.solutionID}.txt fitness{self.solutionID}.txt')
